ENH: Don't unbox object dtype with future.python_scalars (#66307)

+++ pandas/core/indexes/base.py
@@ -981,10 +981,10 @@ class Index(IndexOpsMixin, PandasObject):
             return tuple(self.__array_wrap__(x) for x in result)
         elif method == "reduce":
             result = lib.item_from_zerodim(result)
-            return maybe_unbox_numpy_scalar(result)
+            return maybe_unbox_numpy_scalar(result, dtype=self.dtype)
         elif is_scalar(result):
             # e.g. matmul
-            return maybe_unbox_numpy_scalar(result)
+            return maybe_unbox_numpy_scalar(result, dtype=self.dtype)
 
         if result.dtype == np.float16:
             result = result.astype(np.float32)
@@ -8053,7 +8053,7 @@ class Index(IndexOpsMixin, PandasObject):
             # quick check
             first = self[0]
             if not isna(first):
-                return maybe_unbox_numpy_scalar(first)
+                return maybe_unbox_numpy_scalar(first, dtype=self.dtype)
 
         if not self._is_multi and self.hasnans:
             # Take advantage of cache
@@ -8064,7 +8064,9 @@ class Index(IndexOpsMixin, PandasObject):
         if not self._is_multi and not isinstance(self._values, np.ndarray):
             return self._values._reduce(name="min", skipna=skipna)
 
-        return maybe_unbox_numpy_scalar(nanops.nanmin(self._values, skipna=skipna))
+        return maybe_unbox_numpy_scalar(
+            nanops.nanmin(self._values, skipna=skipna), dtype=self.dtype
+        )
 
     def max(
         self,
@@ -8126,7 +8128,7 @@ class Index(IndexOpsMixin, PandasObject):
             # quick check
             last = self[-1]
             if not isna(last):
-                return maybe_unbox_numpy_scalar(last)
+                return maybe_unbox_numpy_scalar(last, dtype=self.dtype)
 
         if not self._is_multi and self.hasnans:
             # Take advantage of cache
@@ -8137,7 +8139,9 @@ class Index(IndexOpsMixin, PandasObject):
         if not self._is_multi and not isinstance(self._values, np.ndarray):
             return self._values._reduce(name="max", skipna=skipna)
 
-        return maybe_unbox_numpy_scalar(nanops.nanmax(self._values, skipna=skipna))
+        return maybe_unbox_numpy_scalar(
+            nanops.nanmax(self._values, skipna=skipna), dtype=self.dtype
+        )
 
     # --------------------------------------------------------------------
